requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
cachetools==4.2.0
certifi==2020.12.5
chardet==3.0.4
discord.py==1.7.3
httplib2==0.18.1
idna==2.10
multidict==4.7.6
mysql-connector-python==8.0.22
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.4
requests==2.25.1
rsa==4.6
setproctitle==1.2.1
six==1.15.0
soupsieve==2.1
uritemplate==3.0.1
urllib3==1.26.2
yarl==1.5.1

